requirements.txt
alembic==1.3.2
astroid==2.3.3
blinker==1.4
Click==7.0
coverage==5.0.1
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
rope==0.14.0
six==1.13.0
SQLAlchemy==1.3.12
typed-ast==1.4.0
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
